frontend/requirements.txt
streamlit==1.2.0
requests==2.26.0
Pillow==8.3.2
numpy==1.21.2
matplotlib==3.4.3